pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "usf_fabric_monitoring"
version = "0.3.34"
description = "Microsoft Fabric Monitoring & Governance System with Advanced Analytics, Smart Merge Technology, and Star Schema Builder"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.28,<3",
    "pandas>=2.0,<3",
    "numpy>=1.24,<3",
    "plotly>=5.0,<6",
    "openpyxl>=3.1,<4",
    "aiohttp>=3.9,<4",
    "pyyaml>=6.0,<7",
    "jsonschema>=4.17,<5",
    "rich>=13.0,<14",
    "typer>=0.9,<1",
    "python-dotenv>=1.0,<2",
    "azure-identity>=1.15,<2",
    "azure-storage-blob>=12.19,<13",
    "asyncio-throttle>=1.0,<2",
    "pyarrow>=14.0,<18",
    "fastapi>=0.109,<1",
    "uvicorn>=0.27,<1",
    "neo4j>=5.15,<6"
]

[project.scripts]
usf-monitor-hub = "usf_fabric_monitoring.scripts.monitor_hub_pipeline:main"
usf-enforce-access = "usf_fabric_monitoring.scripts.enforce_workspace_access:main"
usf-validate-config = "usf_fabric_monitoring.scripts.validate_config:main"
usf-star-schema = "usf_fabric_monitoring.scripts.build_star_schema:main"
usf-extract-lineage = "usf_fabric_monitoring.scripts.extract_lineage:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0,<9",
    "pytest-cov>=4.0,<6",
    "pytest-asyncio>=0.21,<1",
    "ruff>=0.4,<1",
    "mypy>=1.8,<2",
    "jupyter>=1.0,<2",
    "notebook>=7.0,<8",
    "pyspark>=3.5,<4",
    "delta-spark>=3.0,<4",
    "matplotlib>=3.8,<4",
    "seaborn>=0.13,<1"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires Fabric/Azure credentials and may call real APIs",
]
addopts = "-m 'not integration'"

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
    "B905",   # zip without strict
]

[tool.ruff.lint.isort]
known-first-party = ["usf_fabric_monitoring"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_ignores = true
ignore_missing_imports = true
exclude = ["tests/", "scripts/", "notebooks/"]

